+++ siibra/attributes/__init__.py
@@ -17,4 +17,4 @@ from .attribute import Attribute
 from .attribute_collection import AttributeCollection
 from . import descriptions
 from . import locations
-from . import dataproviders
+from . import datarecipes
