+++ siibra/factory/__init__.py
@@ -21,7 +21,7 @@ from .iterator import iter_preconfigured_ac
 from .livequery import iter_livequery_clss
 
 # import convenient image builders
-from ..attributes.dataproviders.volume.image import (
+from ..attributes.datarecipes.volume.image import (
     from_nifti as imageprovider_from_nifti,
     from_pointcloud as imageprovider_from_pointcloud,
     from_array as imageprovider_from_array,
